+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [Version - 0]
-## improve [v0.34.6] 2023-06-21
+## improve [v0.34.7] 2023-06-21
 - added soring in point menu
 - connected api create & edit & delete qtota
+## improve [v0.34.6] 2023-06-19
+- enhance news menu
+- added category news & challenge
 ## improve [v0.34.5] 2023-06-16
 - enhanced quota report
 ## improve [v0.34.4] 2023-06-16